Fix/refactor

+++ src/main/java/com/example/FullstackProject/service/impl/CategoryServiceImpl.java
@@ -1,57 +1,72 @@
 package com.example.FullstackProject.service.impl;
 
-import com.example.FullstackProject.model.entity.Category;
-import com.example.FullstackProject.model.entity.Product;
+import com.example.FullstackProject.Converter.Converter;
+import com.example.FullstackProject.model.entity.CategoryEntity;
+import com.example.FullstackProject.model.entity.ProductEntity;
+import com.example.FullstackProject.model.request.CategoryRequest;
+import com.example.FullstackProject.model.response.CategoryResponse;
+import com.example.FullstackProject.model.response.ProductResponse;
 import com.example.FullstackProject.repository.CategoryRepository;
 import com.example.FullstackProject.repository.ProductRepository;
 import com.example.FullstackProject.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
 @Transactional
 public class CategoryServiceImpl implements CategoryService {
     private final  CategoryRepository categoryRepository;
+    private final ProductRepository productRepository;
+
     @Override
-    public List<Category> getAllCategories() {
-        return categoryRepository.findAll();
+    public List<CategoryResponse> getAllCategories() {
+        List<CategoryEntity> categoryEntities = categoryRepository.findAll();
+        return Converter.toList(categoryEntities, CategoryResponse.class);
+    }
+
+    public CategoryResponse addCategory(CategoryRequest categoryRequest) {
+        CategoryEntity categoryEntity = Converter.toModel(categoryRequest, CategoryEntity.class);
+        categoryRepository.save(categoryEntity);
+        return Converter.toModel(categoryEntity, CategoryResponse.class);
+
     }
 
     @Override
-    public void addCategory(Category category) {
-        categoryRepository.save(category);
-    }
-
-    @Override
-    public Optional<Category> getCategoryById(Long categoryID) {
-;          Optional<Category> category = Optional.ofNullable(categoryRepository.findById(categoryID).orElseThrow(() -> new IllegalStateException(
-                "category with id " + " does not exist"
-        )));
-        return category;
-    }
-
-    @Override
-    public void deleteCategory(Long categoryID) {
-        categoryRepository.deleteById(categoryID);
-    }
-
-    @Override
-    public void updateCategory(Category categoryDetail) {
-        Category category = categoryRepository.findById(categoryDetail.getId()).orElseThrow(() -> new IllegalStateException(
-                "category with id " + categoryDetail.getId() + " does not exist"
+    public CategoryResponse getCategoryById(Long categoryID) {
+;          CategoryEntity categoryEntity = categoryRepository.findById(categoryID).orElseThrow(() -> new IllegalStateException(
+            "category with id" + " does not exist"
         ));
-        System.out.println(category);
-        System.out.println("categoryDetail");
-        System.out.println( categoryDetail);
-        if (categoryDetail.getName() != null && categoryDetail.getName().length() > 0 && !Objects.equals(category.getName(), categoryDetail.getName())) {
-            category.setName(categoryDetail.getName());
-        }
+        return Converter.toModel(categoryEntity, CategoryResponse.class);
+    }
 
+    @Override
+    public void deleteCategory(Long categoryID) throws Exception {
+
+      ProductEntity productEntity =  productRepository.findAll().stream().filter(
+               pro ->
+                   Objects.equals(pro.getCategory(), categoryID))
+                    .findFirst()
+                    .orElse(null);
+                   if(null != productEntity){
+                        throw new Exception("product is existed");
+                   } else {
+                       categoryRepository.deleteById(categoryID);
+                   }
+    }
+
+    @Override
+    public CategoryResponse updateCategory(CategoryRequest categoryRequest, Long id) {
+       CategoryEntity categoryEntity = categoryRepository.findById(id)
+               .map(category -> {
+                   category.setName(categoryRequest.getName());
+                   return categoryRepository.save(category);
+               }).orElseThrow(() -> new IllegalStateException(
+                   "product with id " + id + " does not exist"
+           ));
+    return Converter.toModel(categoryEntity, CategoryResponse.class);
     }
 }
